refactor: globals and duplicated code
Consolidate global access to an oclif settings file. This also moves the columns to the oclif object which is a breaking change with oclif@1.

Consolidate the screen code into a single file, rather in each package as it was in the old oclif packages.
