--- src/errors/global.d.ts
@@ -1,10 +1,0 @@
-// tslint:disable
-
-declare namespace NodeJS {
-  interface Global {
-    oclif?: {
-      debug?: boolean;
-      errlog?: string;
-    };
-  }
-}
