--- src/help/global.d.ts
@@ -1,5 +1,0 @@
-declare namespace NodeJS {
-  interface Global {
-    'columns': number;
-  }
-}
