+++ test/help/format-root.test.ts
@@ -5,7 +5,7 @@ import {Help} from '../../src/help'
 import {Interfaces} from '../../src'
 
 const g: any = global
-g.columns = 80
+g.oclif.columns = 80
 
 const VERSION = require('../../package.json').version
 const UA = `@oclif/core/${VERSION} ${process.platform}-${process.arch} node-${process.version}`
