+++ test/helpers/init.js
@@ -2,4 +2,5 @@ const path = require('path')
 process.env.TS_NODE_PROJECT = path.resolve('test/tsconfig.json')
 process.env.NODE_ENV = 'development'
 
-global.columns = 80
+global.oclif = global.oclif || {}
+global.oclif.columns = 80
